+++ src/php/FrontendBundle/Domain/StreamHandlerSupplier.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Frontastic\Catwalk\FrontendBundle\Domain;
+
+interface StreamHandlerSupplier
+{
+    public function get(string $streamType): StreamHandlerV2;
+}